pyproject.toml
[tool.poetry]
name = "repl_python3_CSMatchmaker"
version = "0.1.0"
description = ""
authors = ["seanchencs <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
replit = "^3.0.0"
trueskill = "^0.4.5"
discord = "^1.0.1"
flask = "^1.1.2"
discord-py-slash-command = "^1.2.2"
pytz = "^2021.1"
asciichartpy = "^1.5.25"
python-dotenv = "^0.17.1"
tabulate = "^0.8.9"
sqlitedict = "1.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
